Refactor model classes into their own modules
