+++ backend/server/models/__init__.py
@@ -1,0 +1,5 @@
+from .match import Match
+from .ml_model import MLModel
+from .prediction import Prediction
+from .team_match import TeamMatch
+from .team import Team
